@@ -19,11 +19,10 @@ classifiers =
   Operating System :: POSIX
   Operating System :: POSIX :: Linux
   Operating System :: Microsoft :: Windows
-  Programming Language :: Python :: 2.7
-  Programming Language :: Python :: 3.5
   Programming Language :: Python :: 3.6
   Programming Language :: Python :: 3.7
   Programming Language :: Python :: 3.8
+  Programming Language :: Python :: 3.9
   Topic :: Software Development :: Libraries
   Topic :: Multimedia :: Sound/Audio :: Speech
 license_file = LICENSE
@@ -31,11 +30,10 @@ long_description = file: README.md
 long_description_content_type = text/markdown
 
 [options]
-python_requires = >= 2.7
+python_requires = >= 3.6
 include_package_data = True
 packages = find:
 install_requires =
-  six
   beautifulsoup4
   click
   requests
@@ -48,7 +46,6 @@ tests =
   flake8
   testfixtures
   mock
-  six
 docs =
   sphinx
   sphinx-autobuild
